# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Soding Assignment"
  },
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "tags": [
    {
      "name": "api"
    }
  ],
  "paths": {
    "/create": {
      "post": {
        "tags": [
          "api"
        ],
        "operationId": "createTask",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "assignee",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "reporter",
            "in": "formData",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {}
          }
        }
      }
    },
    "/get": {
      "get": {
        "tags": [
          "api"
        ],
        "operationId": "getTask",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {}
          }
        }
      }
    },
    "/list": {
      "get": {
        "tags": [
          "api"
        ],
        "operationId": "listAll",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {}
          }
        }
      }
    },
    "/update": {
      "post": {
        "tags": [
          "api"
        ],
        "operationId": "updateTask",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "assignee",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "reporter",
            "in": "formData",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {}
          }
        }
      }
    },
    "/delete": {
      "get": {
        "tags": [
          "api"
        ],
        "operationId": "updateTask",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {}
          }
        }
      }
    }
  }
}